fix(lens): extract createPerTerminalVisibility shared by find bar and pager
useHistoryPager and useTerminalSearch were near-clones of the same
per-terminal toggleable-overlay receptacle; hoist the store +
close-on-active-switch effect + six verbs into createPerTerminalVisibility,
and define both hooks as createSharedRoot(() => createPerTerminalVisibility())
so each keeps its own isolated singleton.

Agreed by the lowy ⇄ hickey lens debate (finding lowy-8, raised by lowy). Not pushed or merged.
